shut_the_box: Repurpose as game entrypoint
Moved player interaction code out of `game.py` and into
`shut_the_box.py`.
